Reflection of xor-in when reflect_out is set
The xor-in value must be reflected whenever the reflect-out parameter
is set *not* the reflect-in: reflect-in only applies to individual
bytes in the input, reflect-out applies to the reflection of the
register. The reflection of xor-in was missing for strings (and
hex-strings) and was wrongly applied when reflect-in was set (not
reflect-out) for files.